refactor: Rework population middleware
